# python-client-generated/swagger_client/models/driver_single_data.py
# coding: utf-8

"""
    Taxi Aggregator DB Service

    обращения операторов Такси к БД водителей.  # noqa: E501

    OpenAPI spec version: 1.0
    
    Generated by: https://github.com/swagger-api/swagger-codegen.git
"""

import pprint
import re  # noqa: F401

import six

class DriverSingleData(object):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    """
    Attributes:
      swagger_types (dict): The key is attribute name
                            and the value is attribute type.
      attribute_map (dict): The key is attribute name
                            and the value is json key in definition.
    """
    swagger_types = {
        'hash_id': 'DriverHash',
        'online': 'SingleData'
    }

    attribute_map = {
        'hash_id': 'hash_id',
        'online': 'online'
    }

    def __init__(self, hash_id=None, online=None):  # noqa: E501
        """DriverSingleData - a model defined in Swagger"""  # noqa: E501
        self._hash_id = None
        self._online = None
        self.discriminator = None
        self.hash_id = hash_id
        self.online = online

    @property
    def hash_id(self):
        """Gets the hash_id of this DriverSingleData.  # noqa: E501


        :return: The hash_id of this DriverSingleData.  # noqa: E501
        :rtype: DriverHash
        """
        return self._hash_id

    @hash_id.setter
    def hash_id(self, hash_id):
        """Sets the hash_id of this DriverSingleData.


        :param hash_id: The hash_id of this DriverSingleData.  # noqa: E501
        :type: DriverHash
        """
        if hash_id is None:
            raise ValueError("Invalid value for `hash_id`, must not be `None`")  # noqa: E501

        self._hash_id = hash_id

    @property
    def online(self):
        """Gets the online of this DriverSingleData.  # noqa: E501


        :return: The online of this DriverSingleData.  # noqa: E501
        :rtype: SingleData
        """
        return self._online

    @online.setter
    def online(self, online):
        """Sets the online of this DriverSingleData.


        :param online: The online of this DriverSingleData.  # noqa: E501
        :type: SingleData
        """
        if online is None:
            raise ValueError("Invalid value for `online`, must not be `None`")  # noqa: E501

        self._online = online

    def to_dict(self):
        """Returns the model properties as a dict"""
        result = {}

        for attr, _ in six.iteritems(self.swagger_types):
            value = getattr(self, attr)
            if isinstance(value, list):
                result[attr] = list(map(
                    lambda x: x.to_dict() if hasattr(x, "to_dict") else x,
                    value
                ))
            elif hasattr(value, "to_dict"):
                result[attr] = value.to_dict()
            elif isinstance(value, dict):
                result[attr] = dict(map(
                    lambda item: (item[0], item[1].to_dict())
                    if hasattr(item[1], "to_dict") else item,
                    value.items()
                ))
            else:
                result[attr] = value
        if issubclass(DriverSingleData, dict):
            for key, value in self.items():
                result[key] = value

        return result

    def to_str(self):
        """Returns the string representation of the model"""
        return pprint.pformat(self.to_dict())

    def __repr__(self):
        """For `print` and `pprint`"""
        return self.to_str()

    def __eq__(self, other):
        """Returns true if both objects are equal"""
        if not isinstance(other, DriverSingleData):
            return False

        return self.__dict__ == other.__dict__

    def __ne__(self, other):
        """Returns true if both objects are not equal"""
        return not self == other