# python-client-generated/swagger_client/models/hour_response.py
# coding: utf-8

"""
    Taxi Aggregator DB Service

    обращения операторов Такси к БД водителей.  # noqa: E501

    OpenAPI spec version: 1.0
    
    Generated by: https://github.com/swagger-api/swagger-codegen.git
"""

import pprint
import re  # noqa: F401

import six

class HourResponse(object):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    """
    Attributes:
      swagger_types (dict): The key is attribute name
                            and the value is attribute type.
      attribute_map (dict): The key is attribute name
                            and the value is json key in definition.
    """
    swagger_types = {
        'drivers': 'list[DriverSingleData]'
    }

    attribute_map = {
        'drivers': 'drivers'
    }

    def __init__(self, drivers=None):  # noqa: E501
        """HourResponse - a model defined in Swagger"""  # noqa: E501
        self._drivers = None
        self.discriminator = None
        if drivers is not None:
            self.drivers = drivers

    @property
    def drivers(self):
        """Gets the drivers of this HourResponse.  # noqa: E501

        массив структур вида [{hash_id, online}, ...]  # noqa: E501

        :return: The drivers of this HourResponse.  # noqa: E501
        :rtype: list[DriverSingleData]
        """
        return self._drivers

    @drivers.setter
    def drivers(self, drivers):
        """Sets the drivers of this HourResponse.

        массив структур вида [{hash_id, online}, ...]  # noqa: E501

        :param drivers: The drivers of this HourResponse.  # noqa: E501
        :type: list[DriverSingleData]
        """

        self._drivers = drivers

    def to_dict(self):
        """Returns the model properties as a dict"""
        result = {}

        for attr, _ in six.iteritems(self.swagger_types):
            value = getattr(self, attr)
            if isinstance(value, list):
                result[attr] = list(map(
                    lambda x: x.to_dict() if hasattr(x, "to_dict") else x,
                    value
                ))
            elif hasattr(value, "to_dict"):
                result[attr] = value.to_dict()
            elif isinstance(value, dict):
                result[attr] = dict(map(
                    lambda item: (item[0], item[1].to_dict())
                    if hasattr(item[1], "to_dict") else item,
                    value.items()
                ))
            else:
                result[attr] = value
        if issubclass(HourResponse, dict):
            for key, value in self.items():
                result[key] = value

        return result

    def to_str(self):
        """Returns the string representation of the model"""
        return pprint.pformat(self.to_dict())

    def __repr__(self):
        """For `print` and `pprint`"""
        return self.to_str()

    def __eq__(self, other):
        """Returns true if both objects are equal"""
        if not isinstance(other, HourResponse):
            return False

        return self.__dict__ == other.__dict__

    def __ne__(self, other):
        """Returns true if both objects are not equal"""
        return not self == other